Remove webform_node as direct dependency.
In many use cases the webform build will depend only on D8 webform
config entities for build and view the webforms so embedding those forms
on nodes is not always essential in some use cases. Direct dependency
force to install webform_node when not desired. As migrations don't
directly depend on API of this module and there is a conditional that
only references webforms to nodes if the webform node exists is likely
that there is not real harm to avoid unwanted installation.